Save snapshots to temporary directory
The app currently takes a flag/var with a directory where the temporary
images (screenshots) and the output gif is to be saved. This is error
prone as there could be other images in that directory and the glob
command is not using such a strict pattern. In addition, unrelated
images could be deleted.

* replace the mentioned flag with one called `outputDir` that is where
  the output gif file will be saved and defaults to the current
directory.

* create a temporary directory where to keep the snapshots.

* replace the cleaning up function that was looging for a pattern
  in the filenames in the img saving directory with one to delete the
temporary directory.

* reduce the default initial sleep seconds and animation duration.