Add resizing functionality
Add the ability to rescale the screenshots before the animation. When
using smaller images, the (gif) file size is significantly reduced and so
is the time elapsed while producing the animation.

Add a new flag/env variable to set the desired width of the output gif
in pixels. If it's set to a non-zero value, resize each capture before
encoding the image into the file.