Clean up temp images on cancel signal
The clean up function is called if the screenshooting loop ends either
because it's done or because it errors, however if a termination signal
is caught, the temp files are not cleaned.

Add the `imgSaveDir` field to `config`, move the temporary directory
creation and the call to clean up files out of the `run` and make the
call after run finishes or is terminated.